Make component for small Cards with new prop called setUser
